+++ project/src/recoil/setOverlay.ts
@@ -1,0 +1,8 @@
+import { atom } from "recoil";
+
+
+export const setOverlay = atom<boolean>({
+  key: "setOverlay",
+  default: false,
+});
+
